Task 900 (replacing report to history tab)

+++ database/migrations/2015_05_05_101252_create_readings_table.php
@@ -19,7 +19,7 @@ class CreateReadingsTable extends Migration {
 			$table->string('microchip_id')->nullable();
 			$table->double('temperature', 15, 8)->nullable();
 			$table->dateTime('reading_time')->nullable();
-			$table->integer('device_id')->unsigned()->nullable();
+			$table->bigInteger('device_id')->unsigned()->nullable();
 			$table->integer('pet_id')->unsigned()->nullable();
 			$table->timestamps();
 			$table->softDeletes();
